+++ Materials/GodotSea.tres
@@ -1,0 +1,17 @@
+[gd_resource type="ShaderMaterial" load_steps=4 format=3 uid="uid://ctt61l6xpe7hg"]
+
+[ext_resource type="Shader" uid="uid://bx4btogx7mwo0" path="res://Shaders/SeaShader.tres" id="1_rtib3"]
+[ext_resource type="Texture2D" uid="uid://p2yerrtjjfkl" path="res://icon.svg" id="2_wdfkc"]
+[ext_resource type="Texture2D" uid="uid://7qs46g5xvmnd" path="res://Textures/Sand_004_Normal.png" id="3_f130s"]
+
+[resource]
+render_priority = 0
+shader = ExtResource("1_rtib3")
+shader_parameter/displace = 1.0
+shader_parameter/UV_Size = Vector2(200, 200)
+shader_parameter/foamColor = Color(0.109804, 0.301961, 0.52549, 1)
+shader_parameter/DisplacePanningSpeed = 0.2
+shader_parameter/DisplaceAmount = 0.095
+shader_parameter/roughness = 0.1
+shader_parameter/DisplaceTexture = ExtResource("3_f130s")
+shader_parameter/AlbedoTexture = ExtResource("2_wdfkc")
