+++ Materials/glass.tres
@@ -4,5 +4,6 @@
 transparency = 1
 albedo_color = Color(0.34902, 0.505882, 0.662745, 0.454902)
 roughness = 0.25
+rim_enabled = true
 backlight = Color(0.752941, 0.909804, 0.909804, 1)
 refraction_enabled = true
