+++ Materials/gridbox-prototype/prototype_512x512_blue.tres
@@ -1,0 +1,8 @@
+[gd_resource type="StandardMaterial3D" load_steps=2 format=3 uid="uid://cctisxflce1o5"]
+
+[ext_resource type="Texture2D" uid="uid://chb65y16yotcy" path="res://Textures/gridbox-prototype/prototype_512x512_blue1.png" id="1_kslub"]
+
+[resource]
+albedo_texture = ExtResource("1_kslub")
+uv1_triplanar = true
+uv1_world_triplanar = true
