--- Materials/gridbox-prototype/prototype_512x512_blue2.tres
@@ -1,8 +1,0 @@
-[gd_resource type="StandardMaterial3D" load_steps=2 format=3 uid="uid://brkqlb8r7dwp0"]
-
-[ext_resource type="Texture2D" uid="uid://crworx7ak41ni" path="res://Textures/gridbox-prototype/prototype_512x512_blue2.png" id="1_fkkmg"]
-
-[resource]
-albedo_texture = ExtResource("1_fkkmg")
-uv1_triplanar = true
-uv1_world_triplanar = true
