--- Materials/gridbox-prototype/prototype_512x512_blue3.tres
@@ -1,8 +1,0 @@
-[gd_resource type="StandardMaterial3D" load_steps=2 format=3 uid="uid://brkqlb8r7dwp0"]
-
-[ext_resource type="Texture2D" uid="uid://1s8gk2pnms3q" path="res://Textures/gridbox-prototype/prototype_512x512_blue3.png" id="1_e4fhp"]
-
-[resource]
-albedo_texture = ExtResource("1_e4fhp")
-uv1_triplanar = true
-uv1_world_triplanar = true
