+++ Materials/gridbox-prototype/prototype_512x512_green.tres
@@ -1,0 +1,8 @@
+[gd_resource type="StandardMaterial3D" load_steps=2 format=3 uid="uid://dpitjbosqveim"]
+
+[ext_resource type="Texture2D" uid="uid://ffy4bbt533rm" path="res://Textures/gridbox-prototype/prototype_512x512_green1.png" id="1_k20ov"]
+
+[resource]
+albedo_texture = ExtResource("1_k20ov")
+uv1_triplanar = true
+uv1_world_triplanar = true
