+++ Materials/gridbox-prototype/prototype_512x512_grey1.tres
@@ -1,4 +1,4 @@
-[gd_resource type="StandardMaterial3D" load_steps=2 format=3 uid="uid://brkqlb8r7dwp0"]
+[gd_resource type="StandardMaterial3D" load_steps=2 format=3 uid="uid://cnwtdhkhhen8h"]
 
 [ext_resource type="Texture2D" uid="uid://bunst11h41iyv" path="res://Textures/gridbox-prototype/prototype_512x512_grey1.png" id="1_3fg4v"]
