--- Materials/gridbox-prototype/prototype_512x512_grey2.tres
@@ -1,8 +1,0 @@
-[gd_resource type="StandardMaterial3D" load_steps=2 format=3 uid="uid://brkqlb8r7dwp0"]
-
-[ext_resource type="Texture2D" uid="uid://c4siiq1mrargy" path="res://Textures/gridbox-prototype/prototype_512x512_grey2.png" id="1_lq671"]
-
-[resource]
-albedo_texture = ExtResource("1_lq671")
-uv1_triplanar = true
-uv1_world_triplanar = true
