--- Materials/gridbox-prototype/prototype_512x512_grey3.tres
@@ -1,8 +1,0 @@
-[gd_resource type="StandardMaterial3D" load_steps=2 format=3 uid="uid://brkqlb8r7dwp0"]
-
-[ext_resource type="Texture2D" uid="uid://dnlgipv3l75a4" path="res://Textures/gridbox-prototype/prototype_512x512_grey3.png" id="1_1pqoi"]
-
-[resource]
-albedo_texture = ExtResource("1_1pqoi")
-uv1_triplanar = true
-uv1_world_triplanar = true
