+++ Materials/gridbox-prototype/prototype_512x512_grey4.tres
@@ -1,8 +1,8 @@
-[gd_resource type="StandardMaterial3D" load_steps=2 format=3 uid="uid://brkqlb8r7dwp0"]
+[gd_resource type="StandardMaterial3D" load_steps=2 format=3 uid="uid://bbjoehbnon2kc"]
 
-[ext_resource type="Texture2D" uid="uid://brrltur2swnsw" path="res://Textures/gridbox-prototype/prototype_512x512_grey4.png" id="1_gsunb"]
+[ext_resource type="Texture2D" uid="uid://brrltur2swnsw" path="res://Textures/gridbox-prototype/prototype_512x512_grey4.png" id="1_jgs0d"]
 
 [resource]
-albedo_texture = ExtResource("1_gsunb")
+albedo_texture = ExtResource("1_jgs0d")
 uv1_triplanar = true
 uv1_world_triplanar = true
