--- Materials/gridbox-prototype/prototype_512x512_purple.tres
@@ -1,8 +1,0 @@
-[gd_resource type="StandardMaterial3D" load_steps=2 format=3 uid="uid://brkqlb8r7dwp0"]
-
-[ext_resource type="Texture2D" uid="uid://dkbb15e8mv6dk" path="res://Textures/gridbox-prototype/prototype_512x512_purple.png" id="1_vf4oj"]
-
-[resource]
-albedo_texture = ExtResource("1_vf4oj")
-uv1_triplanar = true
-uv1_world_triplanar = true
