--- Materials/gridbox-prototype/prototype_512x512_yellow.tres
@@ -1,9 +1,0 @@
-[gd_resource type="StandardMaterial3D" load_steps=2 format=3 uid="uid://brkqlb8r7dwp0"]
-
-[sub_resource type="CompressedTexture2D" id="CompressedTexture2D_22j55"]
-load_path = "res://.godot/imported/prototype_512x512_yellow.png-4c21a76ea02b590b03d0f951c87026bf.s3tc.ctex"
-
-[resource]
-albedo_texture = SubResource("CompressedTexture2D_22j55")
-uv1_triplanar = true
-uv1_world_triplanar = true
