+++ Prefabs/goal.tscn
@@ -1,7 +1,8 @@
-[gd_scene load_steps=12 format=3 uid="uid://c30crp15bwp25"]
+[gd_scene load_steps=13 format=3 uid="uid://c30crp15bwp25"]
 
 [ext_resource type="Script" path="res://Scripts/Goal.cs" id="1_3wr42"]
 [ext_resource type="Material" uid="uid://grjijbhlfp5j" path="res://Materials/transparent_red.tres" id="2_rs3qs"]
+[ext_resource type="Material" uid="uid://bbjoehbnon2kc" path="res://Materials/gridbox-prototype/prototype_512x512_grey4.tres" id="3_uk4us"]
 [ext_resource type="Texture2D" uid="uid://p2yerrtjjfkl" path="res://icon.svg" id="3_wsvk6"]
 
 [sub_resource type="BoxShape3D" id="BoxShape3D_6wsvi"]
@@ -9,9 +10,11 @@ size = Vector3(6.0271, 3.35962, 0.623291)
 
 [sub_resource type="BoxMesh" id="BoxMesh_q6inv"]
 
-[sub_resource type="BoxMesh" id="BoxMesh_xg717"]
-
 [sub_resource type="BoxShape3D" id="BoxShape3D_8tnqj"]
+
+[sub_resource type="BoxMesh" id="BoxMesh_xg717"]
+material = ExtResource("3_uk4us")
+size = Vector3(1, 3.44995, 1)
 
 [sub_resource type="StandardMaterial3D" id="StandardMaterial3D_0woqj"]
 transparency = 1
@@ -47,23 +50,27 @@ transform = Transform3D(6.0271, 0, 0, 0, 3.35962, 0, 0, 0, 0.623291, 0, 0, 0)
 material_override = ExtResource("2_rs3qs")
 mesh = SubResource("BoxMesh_q6inv")
 
-[node name="Pillar" type="MeshInstance3D" parent="."]
-transform = Transform3D(1, 0, 0, 0, 3.44995, 0, 0, 0, 1, -3.32762, 0, 0)
-mesh = SubResource("BoxMesh_xg717")
+[node name="StaticBody3D2" type="StaticBody3D" parent="."]
+transform = Transform3D(1, 0, 0, 0, 1, 0, 0, 0, 1, 3.328, 0, 0)
 
-[node name="StaticBody3D" type="StaticBody3D" parent="Pillar"]
-
-[node name="CollisionShape3D" type="CollisionShape3D" parent="Pillar/StaticBody3D"]
+[node name="CollisionShape3D" type="CollisionShape3D" parent="StaticBody3D2"]
 shape = SubResource("BoxShape3D_8tnqj")
 
-[node name="Pillar2" type="MeshInstance3D" parent="."]
-transform = Transform3D(1, 0, 0, 0, 3.44995, 0, 0, 0, 1, 3.328, 0, 0)
+[node name="Pillar2" type="MeshInstance3D" parent="StaticBody3D2"]
+transform = Transform3D(1, 0, 0, 0, 1, 0, 0, 0, 1, 4.76837e-07, 0, 0)
 mesh = SubResource("BoxMesh_xg717")
+skeleton = NodePath("../..")
 
-[node name="StaticBody3D" type="StaticBody3D" parent="Pillar2"]
+[node name="StaticBody3D" type="StaticBody3D" parent="."]
+transform = Transform3D(1, 0, 0, 0, 1, 0, 0, 0, 1, -3.32762, 0, 0)
 
-[node name="CollisionShape3D" type="CollisionShape3D" parent="Pillar2/StaticBody3D"]
+[node name="CollisionShape3D" type="CollisionShape3D" parent="StaticBody3D"]
 shape = SubResource("BoxShape3D_8tnqj")
+
+[node name="Pillar" type="MeshInstance3D" parent="StaticBody3D"]
+transform = Transform3D(1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0)
+mesh = SubResource("BoxMesh_xg717")
+skeleton = NodePath("../..")
 
 [node name="Particles" type="CPUParticles3D" parent="."]
 emitting = false
