+++ Scenes/Levels/02_Level.tscn
@@ -1,11 +1,11 @@
-[gd_scene load_steps=15 format=3 uid="uid://o6mq3phmufrm"]
+[gd_scene load_steps=13 format=3 uid="uid://o6mq3phmufrm"]
 
 [ext_resource type="Script" path="res://Scripts/LevelManager.cs" id="1_qy2b1"]
+[ext_resource type="Material" uid="uid://dpitjbosqveim" path="res://Materials/gridbox-prototype/prototype_512x512_green.tres" id="2_semh0"]
 [ext_resource type="PackedScene" uid="uid://ngddmglqy453" path="res://Prefabs/spawn.tscn" id="3_5s221"]
-[ext_resource type="Texture2D" uid="uid://b1mceqyv2wufm" path="res://Textures/gridbox-prototype/prototype_512x512_brown.png" id="3_h3qle"]
+[ext_resource type="Material" uid="uid://bbjoehbnon2kc" path="res://Materials/gridbox-prototype/prototype_512x512_grey4.tres" id="3_ecu5w"]
 [ext_resource type="PackedScene" uid="uid://c30crp15bwp25" path="res://Prefabs/goal.tscn" id="4_7wtrw"]
 [ext_resource type="PackedScene" uid="uid://drxuym5xt5keq" path="res://Prefabs/godot_sea.tscn" id="4_wjhdo"]
-[ext_resource type="Material" uid="uid://brkqlb8r7dwp0" path="res://Materials/gridbox-prototype/prototype_512x512_yellow.tres" id="4_ws223"]
 
 [sub_resource type="Curve3D" id="Curve3D_q27ab"]
 _data = {
@@ -14,17 +14,13 @@ _data = {
 }
 point_count = 5
 
-[sub_resource type="StandardMaterial3D" id="StandardMaterial3D_0sn4c"]
-albedo_texture = ExtResource("3_h3qle")
-uv1_triplanar = true
-uv1_world_triplanar = true
-
 [sub_resource type="BoxMesh" id="BoxMesh_ry10v"]
+material = ExtResource("3_ecu5w")
 
 [sub_resource type="BoxShape3D" id="BoxShape3D_7rkjr"]
 
 [sub_resource type="BoxMesh" id="BoxMesh_vjh80"]
-material = ExtResource("4_ws223")
+material = ExtResource("3_ecu5w")
 
 [sub_resource type="BoxShape3D" id="BoxShape3D_iugmb"]
 
@@ -34,11 +30,6 @@ _data = {
 "tilts": PackedFloat32Array(0, -0.34181, -0.652789, -0.34246, 0)
 }
 point_count = 5
-
-[sub_resource type="StandardMaterial3D" id="StandardMaterial3D_5fqu3"]
-albedo_texture = ExtResource("3_h3qle")
-uv1_triplanar = true
-uv1_world_triplanar = true
 
 [node name="Level02" type="Node3D"]
 script = ExtResource("1_qy2b1")
@@ -61,7 +52,7 @@ path_local = false
 path_continuous_u = true
 path_u_distance = 1.0
 path_joined = false
-material = SubResource("StandardMaterial3D_0sn4c")
+material = ExtResource("2_semh0")
 
 [node name="MeshInstance3D" type="MeshInstance3D" parent="Terrain"]
 transform = Transform3D(2.92306, 0, 0, 0, 1, 0, 0, 0, 6.44715, 1.34051, 5.68757, 0)
@@ -109,7 +100,7 @@ path_local = false
 path_continuous_u = true
 path_u_distance = 1.0
 path_joined = false
-material = SubResource("StandardMaterial3D_5fqu3")
+material = ExtResource("2_semh0")
 
 [node name="Godot Sea" parent="Terrain" instance=ExtResource("4_wjhdo")]
 transform = Transform3D(1000, 0, 0, 0, 1, 0, 0, 0, 1000, 12, -38.6832, -2)
