--- Shaders/DisplaceShader.tres
@@ -1,178 +1,0 @@
-[gd_resource type="VisualShader" load_steps=17 format=3 uid="uid://bx4btogx7mwo0"]
-
-[sub_resource type="VisualShaderNodeTexture2DParameter" id="VisualShaderNodeTexture2DParameter_mssef"]
-parameter_name = "AlbedoTexture"
-texture_type = 1
-texture_filter = 6
-
-[sub_resource type="VisualShaderNodeTexture" id="VisualShaderNodeTexture_jyepi"]
-source = 5
-
-[sub_resource type="VisualShaderNodeVectorOp" id="VisualShaderNodeVectorOp_cc35t"]
-default_input_values = [0, Vector2(0, 0), 1, Vector2(0, 0)]
-op_type = 0
-operator = 2
-
-[sub_resource type="VisualShaderNodeInput" id="VisualShaderNodeInput_wbfmx"]
-input_name = "time"
-
-[sub_resource type="VisualShaderNodeFloatOp" id="VisualShaderNodeFloatOp_8f5xb"]
-
-[sub_resource type="VisualShaderNodeFloatParameter" id="VisualShaderNodeFloatParameter_1stdu"]
-parameter_name = "displace"
-default_value_enabled = true
-default_value = 1.0
-
-[sub_resource type="VisualShaderNodeFloatOp" id="VisualShaderNodeFloatOp_vndat"]
-operator = 2
-
-[sub_resource type="VisualShaderNodeFloatOp" id="VisualShaderNodeFloatOp_kv08x"]
-default_input_values = [0, 0.0, 1, 1000.0]
-operator = 2
-
-[sub_resource type="VisualShaderNodeFloatFunc" id="VisualShaderNodeFloatFunc_fu47v"]
-function = 0
-
-[sub_resource type="VisualShaderNodeInput" id="VisualShaderNodeInput_e4g3p"]
-input_name = "vertex"
-
-[sub_resource type="VisualShaderNodeVectorDecompose" id="VisualShaderNodeVectorDecompose_vm2yt"]
-
-[sub_resource type="VisualShaderNodeFloatOp" id="VisualShaderNodeFloatOp_nktix"]
-
-[sub_resource type="VisualShaderNodeVectorCompose" id="VisualShaderNodeVectorCompose_nl5t5"]
-
-[sub_resource type="VisualShaderNodeFloatOp" id="VisualShaderNodeFloatOp_0bme2"]
-
-[sub_resource type="VisualShaderNodeVec2Parameter" id="VisualShaderNodeVec2Parameter_mg6dp"]
-parameter_name = "UV_Size"
-default_value_enabled = true
-default_value = Vector2(1, 1)
-
-[sub_resource type="VisualShaderNodeInput" id="VisualShaderNodeInput_qlsck"]
-input_name = "uv"
-
-[resource]
-code = "shader_type spatial;
-render_mode blend_mix, depth_draw_opaque, cull_back, diffuse_lambert, specular_schlick_ggx;
-
-uniform float displace = 1;
-uniform vec2 UV_Size = vec2(1.000000, 1.000000);
-uniform sampler2D AlbedoTexture : source_color, filter_linear_mipmap_anisotropic;
-
-
-
-void vertex() {
-// Input:3
-	vec3 n_out3p0 = VERTEX;
-
-
-// VectorDecompose:4
-	float n_out4p0 = n_out3p0.x;
-	float n_out4p1 = n_out3p0.y;
-	float n_out4p2 = n_out3p0.z;
-
-
-// FloatParameter:13
-	float n_out13p0 = displace;
-
-
-// Input:11
-	float n_out11p0 = TIME;
-
-
-// FloatOp:5
-	float n_out5p0 = n_out4p0 + n_out4p2;
-
-
-// FloatOp:15
-	float n_in15p1 = 1000.00000;
-	float n_out15p0 = n_out5p0 * n_in15p1;
-
-
-// FloatOp:12
-	float n_out12p0 = n_out11p0 + n_out15p0;
-
-
-// FloatFunc:2
-	float n_out2p0 = sin(n_out12p0);
-
-
-// FloatOp:14
-	float n_out14p0 = n_out13p0 * n_out2p0;
-
-
-// FloatOp:7
-	float n_out7p0 = n_out4p1 + n_out14p0;
-
-
-// VectorCompose:6
-	vec3 n_out6p0 = vec3(n_out4p0, n_out7p0, n_out4p2);
-
-
-// Input:9
-	vec2 n_out9p0 = UV;
-
-
-// Vector2Parameter:8
-	vec2 n_out8p0 = UV_Size;
-
-
-// VectorOp:10
-	vec2 n_out10p0 = n_out9p0 * n_out8p0;
-
-
-// Output:0
-	VERTEX = n_out6p0;
-	UV = n_out10p0;
-
-
-}
-
-void fragment() {
-	vec4 n_out4p0;
-// Texture2D:4
-	n_out4p0 = texture(AlbedoTexture, UV);
-
-
-// Output:0
-	ALBEDO = vec3(n_out4p0.xyz);
-
-
-}
-"
-nodes/vertex/0/position = Vector2(1020, 280)
-nodes/vertex/2/node = SubResource("VisualShaderNodeFloatFunc_fu47v")
-nodes/vertex/2/position = Vector2(240, 160)
-nodes/vertex/3/node = SubResource("VisualShaderNodeInput_e4g3p")
-nodes/vertex/3/position = Vector2(-1040, 380)
-nodes/vertex/4/node = SubResource("VisualShaderNodeVectorDecompose_vm2yt")
-nodes/vertex/4/position = Vector2(-740, 360)
-nodes/vertex/5/node = SubResource("VisualShaderNodeFloatOp_nktix")
-nodes/vertex/5/position = Vector2(-500, 160)
-nodes/vertex/6/node = SubResource("VisualShaderNodeVectorCompose_nl5t5")
-nodes/vertex/6/position = Vector2(720, 340)
-nodes/vertex/7/node = SubResource("VisualShaderNodeFloatOp_0bme2")
-nodes/vertex/7/position = Vector2(480, 300)
-nodes/vertex/8/node = SubResource("VisualShaderNodeVec2Parameter_mg6dp")
-nodes/vertex/8/position = Vector2(380, 660)
-nodes/vertex/9/node = SubResource("VisualShaderNodeInput_qlsck")
-nodes/vertex/9/position = Vector2(360, 560)
-nodes/vertex/10/node = SubResource("VisualShaderNodeVectorOp_cc35t")
-nodes/vertex/10/position = Vector2(760, 600)
-nodes/vertex/11/node = SubResource("VisualShaderNodeInput_wbfmx")
-nodes/vertex/11/position = Vector2(-320, 40)
-nodes/vertex/12/node = SubResource("VisualShaderNodeFloatOp_8f5xb")
-nodes/vertex/12/position = Vector2(0, 140)
-nodes/vertex/13/node = SubResource("VisualShaderNodeFloatParameter_1stdu")
-nodes/vertex/13/position = Vector2(160, -40)
-nodes/vertex/14/node = SubResource("VisualShaderNodeFloatOp_vndat")
-nodes/vertex/14/position = Vector2(440, 100)
-nodes/vertex/15/node = SubResource("VisualShaderNodeFloatOp_kv08x")
-nodes/vertex/15/position = Vector2(-340, 160)
-nodes/vertex/connections = PackedInt32Array(3, 0, 4, 0, 4, 0, 5, 0, 4, 2, 6, 2, 4, 1, 7, 0, 7, 0, 6, 1, 6, 0, 0, 0, 9, 0, 10, 0, 8, 0, 10, 1, 10, 0, 0, 4, 11, 0, 12, 0, 12, 0, 2, 0, 13, 0, 14, 0, 2, 0, 14, 1, 14, 0, 7, 1, 4, 0, 6, 0, 5, 0, 15, 0, 15, 0, 12, 1, 4, 2, 5, 1)
-nodes/fragment/3/node = SubResource("VisualShaderNodeTexture2DParameter_mssef")
-nodes/fragment/3/position = Vector2(-540, 260)
-nodes/fragment/4/node = SubResource("VisualShaderNodeTexture_jyepi")
-nodes/fragment/4/position = Vector2(100, 300)
-nodes/fragment/connections = PackedInt32Array(3, 0, 4, 2, 4, 0, 0, 0)
