+++ project.godot
@@ -70,3 +70,4 @@ common/physics_interpolation=true
 textures/default_filters/anisotropic_filtering_level=4
 environment/ssao/quality=3
 environment/ssao/half_size=false
+anti_aliasing/quality/msaa_3d=1
